Untested implementation of buffer and texture copying.

+++ src/NeoDemo/Objects/MaterialPropsAndBuffer.cs
@@ -24,7 +24,7 @@ namespace Veldrid.NeoDemo.Objects
         public void CreateDeviceObjects(GraphicsDevice gd, CommandList cl, SceneContext sc)
         {
             UniformBuffer = gd.ResourceFactory.CreateBuffer(
-                new BufferDescription((ulong)Unsafe.SizeOf<MaterialProperties>(), BufferUsage.UniformBuffer));
+                new BufferDescription((uint)Unsafe.SizeOf<MaterialProperties>(), BufferUsage.UniformBuffer));
             cl.UpdateBuffer(UniformBuffer, 0, ref _properties);
         }
